pyproject.toml
[project]
authors = [{name = "Geoffrey Hutchison", email = "geoff.hutchison@gmail.com"}]
dependencies = ["openbabel-wheel>=3.1.1.23,<4"]
license = "BSD-3-Clause"
name = "avogadro-ob-energy"
requires-python = ">= 3.11"
version = "0.1.0"
description = "MMFF94 and GAFF energies using Open Babel"

[project.scripts]
avogadro-ob-energy = "avogadro_ob_energy:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/avogadro_ob_energy"]

[tool.avogadro]
minimum-avogadro-version = "1.103"

[[tool.avogadro.energy-models]]
identifier = "MMFF94"
model-name = "MMFF94"
description = "Calculate MMFF94 energies and gradients"
input-format = "cml"
support.gradients = true
support.elements = "1, 6-9, 14-17, 35, 53"
support.unit-cell = false
support.ions = true
support.radicals = false

[[tool.avogadro.energy-models]]
identifier = "GAFF"
model-name = "GAFF"
description = "Calculate GAFF energies and gradients"
input-format = "cml"
support.gradients = true
support.elements = "1, 6-9, 14-17, 35, 53"
support.unit-cell = false
support.ions = true
support.radicals = false

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["win-64", "linux-64", "linux-aarch64", "osx-64", "osx-arm64"]

[tool.pixi.pypi-dependencies]
avogadro_ob_energy = { path = ".", editable = true }

[tool.pixi.tasks]
